+++ .gitignore
@@ -4,12 +4,10 @@ logs
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-pnpm-debug.log*
-lerna-debug.log*
 
 node_modules
 dist
-dist-ssr
+.yarn
 *.local
 
 # Editor directories and files
