+++ package.json
@@ -4,9 +4,7 @@
     "dev": "yarn workspaces foreach -pt --from '{frontend,backend}' run dev"
   },
   "workspaces": [
-    "frontend",
-    "backend",
-    "shared"
+    "packages/*"
   ],
   "packageManager": "yarn@4.0.0-rc.39"
 }
