+++ packages/backend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "backend",
+  "name": "@dispense-takehome/backend",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
@@ -10,15 +10,19 @@
     "build": "npx tsc"
   },
   "dependencies": {
+    "@dispense-takehome/common": "workspace:^",
     "dotenv": "^16.0.3",
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "helmet": "^6.0.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.17",
+    "@types/helmet": "^4.0.0",
     "@types/node": "^18.14.1",
     "@typescript-eslint/eslint-plugin": "^5.53.0",
     "@typescript-eslint/parser": "^5.53.0",
     "concurrently": "^7.6.0",
+    "eslint": "^8.34.0",
     "eslint-plugin-prettier": "^4.2.1",
     "nodemon": "^2.0.20",
     "prettier": "^2.8.4",
