+++ packages/common/index.ts
@@ -1,0 +1,9 @@
+export interface ContactBody {
+  name: string;
+  email: string;
+  address?: string;
+}
+
+export interface Contact extends ContactBody {
+  id: string;
+}
