+++ packages/frontend/package.json
@@ -1,6 +1,5 @@
 {
-  "name": "frontend",
-  "private": false,
+  "name": "@dispense-takehome/frontend",
   "version": "0.0.0",
   "type": "module",
   "scripts": {
