--- shared/index.ts
@@ -1,9 +1,0 @@
-export interface UserInfo {
-    name: string
-    email: string
-    address?: string
-  }
-  
-  export interface User extends UserInfo {
-    id: string;
-  }
